# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: RCOS API
  version: 0.1.0
  description: |
    Restaurant Compliance OS — REST API.

    All responses use the envelope `{ success, data, meta? }` on success or
    `{ success: false, error: { code, message: { ar, en }, details? } }` on failure.

    All endpoints under `/companies`, `/brands`, `/branches`, `/users`,
    `/departments`, `/org-levels` require a Bearer access token; tenant is
    derived from the JWT and enforced server-side.
servers:
  - url: http://localhost:3001/api/v1
    description: Local dev
  - url: /api/v1
    description: Same-origin
...
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List users
      responses:
        '200':
          description: Users
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/UserSummary' }
    post:
      tags: [Users]
      summary: Invite / create a user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/UserSummary' }
  /users/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    patch:
      tags: [Users]
      summary: Update a user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserPatch' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/UserSummary' }
    delete:
      tags: [Users]
      summary: Soft-delete a user
      responses:
        '204': { description: Deleted }
...
  /org-levels:
    get:
      tags: [Org Levels]
      responses:
        '200':
          description: Org levels
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/OrgLevel' }
    post:
      tags: [Org Levels]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrgLevelInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/OrgLevel' }
  /org-levels/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    patch:
      tags: [Org Levels]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrgLevelInput' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/OrgLevel' }
    delete:
      tags: [Org Levels]
      responses:
        '204': { description: Deleted }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    UserSummary:
      type: object
      properties:
        id:              { type: string, format: uuid }
        email:           { type: string, format: email }
        phone:           { type: [string, 'null'] }
        fullNameAr:      { type: string }
        fullNameEn:      { type: string }
        preferredLocale: { type: string, enum: [ar, en] }
        status:          { type: string, enum: [active, invited, suspended, deactivated] }
        branchIds:
          type: array
          items: { type: string, format: uuid }
        departmentId:    { type: [string, 'null'], format: uuid }
        avatarUrl:       { type: [string, 'null'] }
        lastLoginAt:     { type: [string, 'null'], format: date-time }
        createdAt:       { type: string, format: date-time }
    UserInput:
      type: object
      required: [email, password, fullNameAr, fullNameEn, roleKeys]
      properties:
        email:           { type: string, format: email }
        password:        { type: string, minLength: 8 }
        fullNameAr:      { type: string }
        fullNameEn:      { type: string }
        phone:           { type: [string, 'null'] }
        preferredLocale: { type: string, enum: [ar, en] }
        branchIds:
          type: array
          items: { type: string, format: uuid }
        departmentId:    { type: [string, 'null'], format: uuid }
        roleKeys:
          type: array
          items: { type: string }
    UserPatch:
      type: object
      properties:
        fullNameAr:      { type: string }
        fullNameEn:      { type: string }
        phone:           { type: [string, 'null'] }
        preferredLocale: { type: string, enum: [ar, en] }
        branchIds:
          type: array
          items: { type: string, format: uuid }
        departmentId:    { type: [string, 'null'], format: uuid }
        status:          { type: string, enum: [active, invited, suspended, deactivated] }
        avatarUrl:       { type: [string, 'null'] }
...
    OrgLevel:
      type: object
      properties:
        id:            { type: string, format: uuid }
        companyId:     { type: string, format: uuid }
        depth:         { type: integer }
        nameAr:        { type: string }
        nameEn:        { type: string }
        parentLevelId: { type: [string, 'null'], format: uuid }
        createdAt:     { type: string, format: date-time }
    OrgLevelInput:
      type: object
      required: [depth, nameAr, nameEn]
      properties:
        depth:         { type: integer, minimum: 0, maximum: 10 }
        nameAr:        { type: string }
        nameEn:        { type: string }
        parentLevelId: { type: [string, 'null'], format: uuid }